Add missing interactions
