+++ src/Prove.tsx
@@ -1,8 +1,5 @@
-import Header from "./global-components/Header";
-
 function Prove() {
   return <div>
-    <Header />
   </div>;
 }
