+++ src/global-components/Header.tsx
@@ -1,11 +1,9 @@
 import NavBarEA from "./NavBarEA";
-import NavBarGeneric from "./NavBarGeneric";
 
 function Header() {
     return (
         <div>
             <NavBarEA />
-            <NavBarGeneric />
         </div>
     );
 }
