--- src/global-components/NavBarGeneric.tsx
@@ -1,9 +1,0 @@
-function NavBarGeneric() {
-    return (
-        <div className="h-14">
-
-        </div>
-    );
-}
-
-export default NavBarGeneric;
